more fixes, changes, guesses

+++ unk-gxm-text-draw.h
@@ -1,2 +1,2 @@
 #pragma once
-float unk_gxm_text_draw(int param_1,int param_2,int param_3,uint param_4,ushort *param_5);
+float unk_gxm_text_draw(int param_1,int param_2,int param_3,float param_4,ushort *param_5);
